created buttoon on map page to redirect to room page
