+++ src/components/ErrorPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ErrorPage = () => {
+    return (
+        <div>404 busy just now</div>
+    )
+}
+
+export default ErrorPage
