+++ src/components/Room.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Room = () =>{
+
+    return(
+        <div>
+            hi!
+        </div>
+    )
+
+}
+
+export default Room;
